fix: Move config analyzer to erpc package
