+++ erpc/config_analyzer.go
@@ -1,4 +1,4 @@
-package main
+package erpc
 
 import (
 	"fmt"
